removed dependency on github.com/stretchr/*
